requirements.txt
PyNaCl

discord.py[voice]

youtube-dl

discord.components

urllib3

